Create destinationDirectory and 'latest' directory automatically
It is no longer necessary to manually create empty destination and 'latest' directories (using .gitkeep) before pushing the first version of a report.

The action will now automatically create the directories it needs and it will not fail anymore, when it tries to delete a 'latest' directory that does yet exist.